feat: merge automatically after a specific number of size-based merges
This places a practical limit on the number of rows that are ingested before we
merge periodically; this would end up being d.batchSize * d.manualMergeFreq.

This is intended to solve the issue where a huge number of rows are updated at
once or there is a near-constant stream of reads (which mean that a time-based
sync is never occurring).  Since we want to ensure that we don't go too long
without merging into the target table and have the stage table bloat, this
ensures that we periodically perform our merge.